extract ObserveEngineImpl into its own file
